+++ MachineLearning.Model/Layer/Snapshot/ILayerSnapshot.cs
@@ -51,7 +51,7 @@ public static class LayerSnapshots
         {
             SimpleLayer simpleLayer => new Simple(simpleLayer.InputNodeCount, simpleLayer.OutputNodeCount),
             StringEmbeddingLayer => new Embedding(),
-            TokenOutputLayer or IEmbedder<string, char> or IEmbedder<double[], int> => Empty,
+            TokenOutputLayer or IEmbedder<string, char> or IEmbedder<float[], int> => Empty,
             _ => throw new NotImplementedException($"No snapshot for {layer} found"),
         };
     }
